+++ stream-access-control/access-control-core/src/test/java/com/backbase/stream/service/UserServiceTest.java
@@ -1,25 +1,24 @@
 package com.backbase.stream.service;
 
 import static com.backbase.stream.test.LambdaAssertions.assertEqualsTo;
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.backbase.dbs.user.api.service.v2.IdentityManagementApi;
 import com.backbase.dbs.user.api.service.v2.UserManagementApi;
-import com.backbase.identity.integration.api.service.v1.RealmApi;
+import com.backbase.identity.integration.api.service.v1.IdentityIntegrationServiceApi;
 import com.backbase.identity.integration.api.service.v1.model.EnhancedUserRepresentation;
+import com.backbase.identity.integration.api.service.v1.model.UserRequestBody;
 import com.backbase.stream.legalentity.model.User;
+import java.util.Collections;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import reactor.core.publisher.Mono;
 
 @ExtendWith(MockitoExtension.class)
@@ -34,28 +33,54 @@ class UserServiceTest {
     private IdentityManagementApi identityManagementApi;
 
     @Mock
-    private RealmApi realmApi;
+    private IdentityIntegrationServiceApi identityIntegrationApi;
 
     @BeforeEach
     void setup() {
-        subject = new UserService(usersApi, identityManagementApi, Optional.of(realmApi));
+        subject = new UserService(usersApi, identityManagementApi, Optional.of(identityIntegrationApi));
     }
 
     @Test
-    void lockUser() {
+    void changeEnableStatusToDisable() {
         String internalId = "someInternalId";
         String realm = "someRealm";
         String email = "some@email.com";
 
-        EnhancedUserRepresentation eur = new EnhancedUserRepresentation().id(internalId).email(email);
-        when(realmApi.getUser(internalId, realm)).thenReturn(Mono.just(eur));
-        when(realmApi.putUser(eq(internalId), eq(realm), any())).thenReturn(Mono.empty().then());
+        EnhancedUserRepresentation eur = new EnhancedUserRepresentation().id(internalId).email(email).enabled(true);
+        when(identityIntegrationApi.getUserById(realm, internalId)).thenReturn(Mono.just(eur));
+        when(identityIntegrationApi.updateUserById(eq(realm), eq(internalId), any())).thenReturn(Mono.empty().then());
 
 
-        User user = new User().internalId(internalId);
-        Mono<User> result = subject.lockUser(user, realm);
+        User user = new User().internalId(internalId).locked(true);
+        Mono<User> result = subject.changeEnableStatus(user, realm);
 
 
         result.subscribe(assertEqualsTo(user));
+        verify(identityIntegrationApi).getUserById(eq(realm), eq(internalId));
+        UserRequestBody expectedUser = new UserRequestBody().id(internalId).email(email).enabled(false)
+            .credentials(Collections.emptyList());
+        verify(identityIntegrationApi).updateUserById(eq(realm), eq(internalId), eq(expectedUser));
+    }
+
+    @Test
+    void changeEnableStatusToEnabled() {
+        String internalId = "someInternalId";
+        String realm = "someRealm";
+        String email = "some@email.com";
+
+        EnhancedUserRepresentation eur = new EnhancedUserRepresentation().id(internalId).email(email).enabled(false);
+        when(identityIntegrationApi.getUserById(realm, internalId)).thenReturn(Mono.just(eur));
+        when(identityIntegrationApi.updateUserById(eq(realm), eq(internalId), any())).thenReturn(Mono.empty().then());
+
+
+        User user = new User().internalId(internalId).locked(false);
+        Mono<User> result = subject.changeEnableStatus(user, realm);
+
+
+        result.subscribe(assertEqualsTo(user));
+        verify(identityIntegrationApi).getUserById(eq(realm), eq(internalId));
+        UserRequestBody expectedUser = new UserRequestBody().id(internalId).email(email).enabled(true)
+            .credentials(Collections.emptyList());
+        verify(identityIntegrationApi).updateUserById(eq(realm), eq(internalId), eq(expectedUser));
     }
 }